Refactor header access patterns
Stop exposing registered fields directly as strucutres - since there are registered fields stored on the HeaderState type as well (to track registered fields which can be key dependent). Access is replaced with accessor structs which allow access to all header fields in both shared and mutable reference forms. The access structs depend on the header state, and so can always return the correct types.

Also refactors header documentation to be in markdown files alongside the source, so that header documentation can be easily re-used without copy and paste.